+++ tests/test_sdr/test_read_announcements.py
@@ -1,13 +1,14 @@
 from unittest.mock import patch, Mock
 import pytest
 import pandas as pd
+import re
 from imf_reader import sdr
 from imf_reader.sdr.read_announcements import (
     read_tsv,
     clean_df,
     format_date,
     get_holdings_and_allocations_data,
-    get_latest_allocations_holdings_date,
+    fetch_latest_allocations_holdings_date,
     fetch_allocations_holdings,
     BASE_URL,
     MAIN_PAGE_URL,
@@ -50,7 +51,7 @@ class TestReadAnnouncements:
     def test_read_tsv_failure(self, mock_read_csv):
         """Test read_tsv raises ValueError on malformed data."""
         mock_read_csv.side_effect = pd.errors.ParserError
-        with pytest.raises(ValueError, match="SDR _data not available for this date"):
+        with pytest.raises(ValueError, match="SDR data not available for this date"):
             read_tsv("mock_url")
 
     def test_clean_df_correct_format(self, input_df):
@@ -128,11 +129,11 @@ class TestReadAnnouncements:
     def test_get_holdings_and_allocations_data_failure(self, mock_read_tsv):
         """Test get_holdings_and_allocations_data raises ValueError when read_tsv fails."""
         with pytest.raises(ValueError, match="Data not available"):
-            get_holdings_and_allocations_data(2024, 2)
+            get_holdings_and_allocations_data(1800, 1)
 
     @patch("imf_reader.sdr.read_announcements.make_request")
     def test_get_latest_date_success(self, mock_make_request):
-        """Test get_latest_allocations_holdings_date successfully returns the latest date."""
+        """Test fetch_latest_allocations_holdings_date successfully returns the latest date."""
 
         # Mock HTML content
         mock_html_content = """
@@ -159,7 +160,7 @@ class TestReadAnnouncements:
         mock_make_request.return_value = mock_response
 
         # Call the function
-        result = get_latest_allocations_holdings_date()
+        result = fetch_latest_allocations_holdings_date()
 
         # Assertions
         mock_make_request.assert_called_once_with(MAIN_PAGE_URL)
@@ -167,7 +168,7 @@ class TestReadAnnouncements:
 
     @patch("imf_reader.sdr.read_announcements.make_request")
     def test_get_latest_date_invalid_html(self, mock_make_request):
-        """Test get_latest_allocations_holdings_date raises an error when HTML parsing fails."""
+        """Test fetch_latest_allocations_holdings_date raises an error when HTML parsing fails."""
 
         # Mock malformed HTML content
         mock_response = Mock()
@@ -176,9 +177,9 @@ class TestReadAnnouncements:
 
         # Call the function and expect an IndexError
         with pytest.raises(IndexError):
-            get_latest_allocations_holdings_date()
+            fetch_latest_allocations_holdings_date()
 
-    @patch("imf_reader.sdr.read_announcements.get_latest_allocations_holdings_date")
+    @patch("imf_reader.sdr.read_announcements.fetch_latest_allocations_holdings_date")
     @patch("imf_reader.sdr.read_announcements.clean_df")
     @patch("imf_reader.sdr.read_announcements.read_tsv")
     @patch("imf_reader.sdr.read_announcements.get_holdings_and_allocations_data")
@@ -191,8 +192,8 @@ class TestReadAnnouncements:
         input_df,
     ):
         """Test fetch_allocations_holdings when no date is provided."""
-        # Mock get_latest_allocations_holdings_date to return a specific date
-        mock_get_latest_date.return_value = (2, 2024)
+        # Mock fetch_latest_allocations_holdings_date to return a specific date
+        mock_get_latest_date.return_value = (2024, 2)
 
         # Mock read_tsv
         mock_read_tsv.return_value = input_df
@@ -215,26 +216,28 @@ class TestReadAnnouncements:
         result = fetch_allocations_holdings()
 
         # Assertions
-        mock_get_latest_date.assert_called_once()  # Ensure get_latest_allocations_holdings_date was called
+        mock_get_latest_date.assert_called_once()  # Ensure fetch_latest_allocations_holdings_date was called
         mock_get_holdings_and_allocations_data.assert_called_once_with(
-            2, 2024
+            2024, 2
         )  # Ensure the correct call
         pd.testing.assert_frame_equal(
             result, cleaned_df
         )  # Ensure the result matches the cleaned data
 
-    @patch("imf_reader.sdr.read_announcements.get_holdings_and_allocations_data")
-    @patch("imf_reader.sdr.read_announcements.get_latest_allocations_holdings_date")
-    def test_fetch_allocations_holdings_failure(
-        self, mock_get_latest_date, mock_get_holdings_data
+
+@patch("imf_reader.sdr.read_announcements.get_holdings_and_allocations_data")
+@patch(
+    "imf_reader.sdr.read_announcements.fetch_latest_allocations_holdings_date",
+    return_value=(2024, 1),
+)
+def test_fetch_allocations_holdings_future_date(
+    mock_get_latest_date, mock_get_holdings_and_allocations_data
+):
+    """Test fetch_allocations_holdings when unavailable future date is provided."""
+    with pytest.raises(
+        ValueError,
+        match=re.escape(
+            "SDR data unavailable for: (2025, 1).\nLatest available: (2024, 1)"
+        ),
     ):
-        """Test fetch_allocations_holdings raises ValueError when data fetching fails."""
-        # Mock get_latest_allocations_holdings_date to return a date
-        mock_get_latest_date.return_value = (2, 2024)
-
-        # Simulate a failure in get_holdings_and_allocations_data
-        mock_get_holdings_data.side_effect = ValueError("Data not available")
-
-        # Assertions
-        with pytest.raises(ValueError, match="Data not available"):
-            fetch_allocations_holdings()
+        fetch_allocations_holdings((2025, 1))